refactor: relocate unit tests to respective module
